move to _safe_split to utils/metaestimators.py
