+++ sklearn/multiclass.py
@@ -48,8 +48,9 @@ from .utils.validation import _num_samples
 from .utils.validation import check_is_fitted
 from .utils.validation import check_X_y
 from .utils.multiclass import (_check_partial_fit_first_call,
-                               check_classification_targets,
-                               _safe_split)
+                               check_classification_targets)
+from .utils.metaestimators import _safe_split
+
 from .externals.joblib import Parallel
 from .externals.joblib import delayed
 from .externals.six.moves import zip as izip
