+++ sklearn/tests/test_multiclass.py
@@ -15,7 +15,8 @@ from sklearn.utils.testing import assert_raise_message
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.multiclass import OneVsOneClassifier
 from sklearn.multiclass import OutputCodeClassifier
-from sklearn.utils.multiclass import check_classification_targets, type_of_target, _safe_split
+from sklearn.utils.multiclass import check_classification_targets, type_of_target
+from sklearn.utils.metaestimators import _safe_split
 from sklearn.utils import shuffle
 
 from sklearn.metrics import precision_score
